refactor(chain): version registry cursor persistence
