+++ packages/agent/src/dkg-agent.ts
@@ -717,7 +717,7 @@ function constructConfiguredChainAdapter(
       minPublisherNativeWei: config.chainConfig.minPublisherNativeWei,
       minPublisherTracWei: config.chainConfig.minPublisherTracWei,
       contextGraphRegistryScanCursorStore: config.contextGraphRegistryScanCursorStore,
-      contextGraphRegistryTipScanCursorStore: config.contextGraphRegistryTipScanCursorStore,
+      contextGraphRegistryScanCursorPersistence: config.contextGraphRegistryScanCursorPersistence,
     };
     const chain = config.chainConfig.adminPrivateKey
       ? new EVMChainAdapter({ ...evmConfigBase, adminPrivateKey: config.chainConfig.adminPrivateKey })
