+++ packages/chain/test/context-graph-registry-scan.unit.test.ts
@@ -1,177 +1,21 @@
 import { describe, it, expect, vi } from 'vitest';
-import { EVMChainAdapter, type EVMAdapterConfig } from '../src/evm-adapter.js';
-import { ContextGraphChainScanPartialError, type ContextGraphChainScanOptions, type ContextGraphOnChain, type ContextGraphRegistryScanOptions } from '../src/chain-adapter.js';
+import { EVMChainAdapter } from '../src/evm-adapter.js';
+import { ContextGraphChainScanPartialError, type ContextGraphChainScanOptions } from '../src/chain-adapter.js';
 import {
   CG_REGISTRY_MAX_SCAN_PAGES,
   CG_REGISTRY_REORG_BUFFER_BLOCKS,
 } from '../src/evm-adapter-base.js';
-
-function recorder<A extends unknown[], R>(impl: (...args: A) => R) {
-  const calls: A[] = [];
-  const fn = (...args: A): R => {
-    calls.push(args);
-    return impl(...args);
-  };
-  return Object.assign(fn, { calls });
-}
-
-// A mutable di-seam double: records every call and runs the current `impl`.
-// `setImpl` swaps the steady-state behaviour (the no-mock analogue of
-// `mockResolvedValue`/`mockReturnValue`); `queueOnce` enqueues one-shot
-// outcomes consumed before the steady-state impl (the analogue of
-// `mockResolvedValueOnce`/`mockRejectedValueOnce`); `reset` clears both the
-// recorded calls and any queued/steady behaviour back to a returns-undefined
-// default (the analogue of `mockReset`); `clear` drops only recorded calls
-// (the analogue of `mockClear`).
-type OnceOutcome<R> = { type: 'return'; value: R } | { type: 'throw'; error: unknown };
-function seam<A extends unknown[], R>(initialImpl: (...args: A) => R) {
-  const calls: A[] = [];
-  const queue: OnceOutcome<R>[] = [];
-  let impl = initialImpl;
-  const fn = (...args: A): R => {
-    calls.push(args);
-    if (queue.length > 0) {
-      const next = queue.shift() as OnceOutcome<R>;
-      if (next.type === 'throw') throw next.error;
-      return next.value;
-    }
-    return impl(...args);
-  };
-  return Object.assign(fn, {
-    calls,
-    setImpl(next: (...args: A) => R) {
-      impl = next;
-    },
-    queueOnce(outcome: OnceOutcome<R>) {
-      queue.push(outcome);
-    },
-    reset() {
-      calls.length = 0;
-      queue.length = 0;
-      impl = (() => undefined as unknown as R) as (...args: A) => R;
-    },
-    clear() {
-      calls.length = 0;
-    },
-  });
-}
-
-const DEPLOYER_PK = '0xac0974bec39a17e36ba4a6b4d238ff944bacb478cbed5efcae784d7bf4f2ff80';
-const ADMIN_PK = '0x5de4111afa1a4b94908f83103eb1f1706367c2e68ca870fc3fb9a804cdab365a';
-const REGISTRY = '0x3333333333333333333333333333333333333333';
-
-class MemoryRegistryScanCursorStore {
-  readonly cursorKeyVersion = 2 as const;
-  readonly values = new Map<string, number>();
-  readonly loads: string[] = [];
-  readonly saves: Array<{ key: string; nextBlock: number }> = [];
-
-  async load(key: { chainId: string; deploymentId: string; registryAddress: string; cursorKind: 'historical' | 'tip' }): Promise<number | undefined> {
-    const encoded = this.key(key);
-    this.loads.push(encoded);
-    return this.values.get(encoded);
-  }
-
-  async save(key: { chainId: string; deploymentId: string; registryAddress: string; cursorKind: 'historical' | 'tip' }, nextBlock: number): Promise<void> {
-    const encoded = this.key(key);
-    this.saves.push({ key: encoded, nextBlock });
-    this.values.set(encoded, nextBlock);
-  }
-
-  private key(key: { chainId: string; deploymentId: string; registryAddress: string; cursorKind: 'historical' | 'tip' }): string {
-    return `${key.chainId}|${key.deploymentId}|${key.cursorKind}|${key.registryAddress.toLowerCase()}`;
-  }
-}
-
-class LegacyRolelessRegistryScanCursorStore {
-  readonly values = new Map<string, number>();
-  readonly loads: string[] = [];
-  readonly saves: Array<{ key: string; nextBlock: number }> = [];
-
-  async load(key: { chainId: string; deploymentId: string; registryAddress: string }): Promise<number | undefined> {
-    const encoded = this.key(key);
-    this.loads.push(encoded);
-    return this.values.get(encoded);
-  }
-
-  async save(
-    key: { chainId: string; deploymentId: string; registryAddress: string },
-    nextBlock: number,
-  ): Promise<void> {
-    const encoded = this.key(key);
-    this.saves.push({ key: encoded, nextBlock });
-    this.values.set(encoded, nextBlock);
-  }
-
-  private key(key: { chainId: string; deploymentId: string; registryAddress: string }): string {
-    return `${key.chainId}|${key.deploymentId}|${key.registryAddress.toLowerCase()}`;
-  }
-}
-
-function minimalConfig(overrides: Partial<EVMAdapterConfig> = {}): EVMAdapterConfig {
-  return {
-    rpcUrl: 'http://127.0.0.1:59998',
-    privateKey: DEPLOYER_PK,
-    adminPrivateKey: ADMIN_PK,
-    hubAddress: '0x0000000000000000000000000000000000000001',
-    chainId: 'evm:31337',
-    staticNetwork: false,
-    ...overrides,
-  };
-}
-
-function makeRegistry(overrides: Record<string, unknown> = {}) {
-  return {
-    getAddress: recorder(async () => REGISTRY),
-    filters: { NameClaimed: recorder(() => 'NameClaimedFilter') },
-    interface: {
-      parseLog: recorder(({ data }: { data: string }) => {
-        if (data === '0x01') {
-          return {
-            name: 'NameClaimed',
-            args: {
-              nameHash: '0xaaa0000000000000000000000000000000000000000000000000000000000001',
-              creator: '0x1111111111111111111111111111111111111111',
-              accessPolicy: 0,
-            },
-          };
-        }
-        return null;
-      }),
-    },
-    queryFilter: seam(async (_filter: unknown, _lo: number, _hi: number) => [] as unknown[]),
-    connect: recorder(() => undefined),
-    ...overrides,
-  } as any;
-}
-
-function makeAdapter(registry: any, head = 0, config: Partial<EVMAdapterConfig> = {}) {
-  const adapter = new EVMChainAdapter(minimalConfig(config));
-  registry.connect = recorder(() => registry);
-  const provider = {
-    getBlockNumber: seam(async () => head),
-    getCode: seam(async (_address: string, block?: number) =>
-      block === undefined || block >= 0 ? '0x6000' : '0x',
-    ),
-  };
-  (adapter as any).contracts = { contextGraphNameRegistry: registry };
-  (adapter as any).initialized = true;
-  (adapter as any).provider = provider;
-  (adapter as any).providers = [provider];
-  return { adapter, provider };
-}
-
-async function collectRegistryScan(
-  adapter: EVMChainAdapter,
-  options: ContextGraphRegistryScanOptions,
-): Promise<ContextGraphOnChain[]> {
-  const results: ContextGraphOnChain[] = [];
-  for await (const page of adapter.scanContextGraphRegistryPages(options)) {
-    results.push(...page.contextGraphs);
-    await page.ack();
-  }
-  return results;
-}
+import {
+  MemoryRegistryScanCursorStore,
+  REGISTRY,
+  collectRegistryScan,
+  makeAdapter,
+  makeRegistry,
+  minimalConfig,
+  recorder,
+  roleAwareRegistryCursorPersistence,
+  seam,
+} from './context-graph-registry-scan-test-support.js';
 
 describe('EVMChainAdapter.listContextGraphsFromChain registry scan', () => {
   it('anchors at the registry deploy block and paginates with the 2,000-block default', async () => {
@@ -243,7 +87,7 @@ describe('EVMChainAdapter.listContextGraphsFromChain registry scan', () => {
     const store = new MemoryRegistryScanCursorStore();
     const registry = makeRegistry();
     const { adapter } = makeAdapter(registry, 4_999, {
-      contextGraphRegistryScanCursorStore: store,
+      contextGraphRegistryScanCursorPersistence: roleAwareRegistryCursorPersistence(store),
     });
     registry.queryFilter.queueOnce({
       type: 'return',
@@ -261,226 +105,6 @@ describe('EVMChainAdapter.listContextGraphsFromChain registry scan', () => {
     expect(partial.failedFromBlock).toBe(2_000);
     expect(partial.failedToBlock).toBe(3_999);
     expect(store.saves.map((s) => s.nextBlock)).toEqual([2_000]);
-  });
-
-  it('probes the current tip independently while a middle historical page remains unavailable', async () => {
-    let head = 4_999;
-    const store = new MemoryRegistryScanCursorStore();
-    const tipGraphBlock = 9_999;
-    const betweenTipProbesGraphBlock = 10_001;
-    const registry = makeRegistry({
-      queryFilter: seam(async (_filter: unknown, lo: number, hi: number) => {
-        if (lo >= 2_000 && lo < 4_000) throw new Error('archive range unavailable');
-        return [tipGraphBlock, betweenTipProbesGraphBlock]
-          .filter((blockNumber) => lo <= blockNumber && blockNumber <= hi)
-          .map((blockNumber) => ({ topics: [], data: '0x01', blockNumber }));
-      }),
-    });
-    const { adapter, provider } = makeAdapter(registry, head, {
-      contextGraphRegistryScanCursorStore: store,
-    });
-    provider.getBlockNumber.setImpl(async () => head);
-
-    const firstRecovery = await collectRegistryScan(adapter, {
-      mode: 'seedFull',
-    }).catch((err) => err);
-
-    expect(firstRecovery).toBeInstanceOf(ContextGraphChainScanPartialError);
-    expect((adapter as any).contextGraphRegistryScanCursor.getCachedWatermark(REGISTRY)).toBe(2_000);
-
-    head = 10_000;
-    registry.queryFilter.clear();
-    const tipResults = await collectRegistryScan(adapter, { mode: 'tip' });
-
-    expect(tipResults.map((cg) => cg.blockNumber)).toEqual([tipGraphBlock]);
-    expect(registry.queryFilter.calls.map(([, lo, hi]: [unknown, number, number]) => [lo, hi])).toEqual([
-      [8_001, 10_000],
-    ]);
-    expect((adapter as any).contextGraphRegistryScanCursor.getCachedWatermark(REGISTRY)).toBe(2_000);
-
-    // More than one configured page elapses before the next tip probe. Its independent cursor,
-    // with reorg overlap, closes the interval instead of jumping to only the newest page and
-    // permanently skipping the event immediately after the prior head.
-    head = 13_200;
-    const { adapter: restartedAdapter, provider: restartedProvider } = makeAdapter(registry, head, {
-      contextGraphRegistryScanCursorStore: store,
-    });
-    restartedProvider.getBlockNumber.setImpl(async () => head);
-    registry.queryFilter.clear();
-    const nextTipResults = await collectRegistryScan(restartedAdapter, { mode: 'tip' });
-
-    expect(nextTipResults.map((cg) => cg.blockNumber)).toEqual([
-      tipGraphBlock,
-      betweenTipProbesGraphBlock,
-    ]);
-    expect(registry.queryFilter.calls.map(([, lo, hi]: [unknown, number, number]) => [lo, hi])).toEqual([
-      [9_951, 11_950],
-      [11_951, 13_200],
-    ]);
-    expect((restartedAdapter as any).contextGraphRegistryTipScanCursor.getCachedWatermark(REGISTRY)).toBe(13_201);
-    expect(await restartedAdapter.hasContextGraphRegistryScanWatermark()).toBe(true);
-    expect((restartedAdapter as any).contextGraphRegistryScanCursor.getCachedWatermark(REGISTRY)).toBe(2_000);
-
-    registry.queryFilter.clear();
-    const secondRecovery = await collectRegistryScan(restartedAdapter, {
-      mode: 'seedFull',
-    }).catch((err) => err);
-
-    expect(secondRecovery).toBeInstanceOf(ContextGraphChainScanPartialError);
-    expect(registry.queryFilter.calls.map(([, lo, hi]: [unknown, number, number]) => [lo, hi])).toEqual([
-      [0, 1_999],
-      [2_000, 3_999],
-    ]);
-    expect((restartedAdapter as any).contextGraphRegistryScanCursor.getCachedWatermark(REGISTRY)).toBe(2_000);
-  });
-
-  it('retains the first attempted tip range when its query fails before any acknowledgement', async () => {
-    const store = new MemoryRegistryScanCursorStore();
-    const registry = makeRegistry({
-      queryFilter: seam(async () => {
-        throw new Error('first tip page unavailable');
-      }),
-    });
-    const { adapter } = makeAdapter(registry, 10_000, {
-      contextGraphRegistryScanCursorStore: store,
-    });
-
-    await expect(collectRegistryScan(adapter, { mode: 'tip' }))
-      .rejects.toThrow('first tip page unavailable');
-    expect(store.saves.map((save) => save.nextBlock)).toEqual([8_001]);
-
-    const eventBlock = 10_001;
-    const head = 13_200;
-    registry.queryFilter.reset();
-    registry.queryFilter.setImpl(async (_filter: unknown, lo: number, hi: number) =>
-      lo <= eventBlock && eventBlock <= hi
-        ? [{ topics: [], data: '0x01', blockNumber: eventBlock }]
-        : [],
-    );
-    const { adapter: restarted } = makeAdapter(registry, head, {
-      contextGraphRegistryScanCursorStore: store,
-    });
-
-    const recovered = await collectRegistryScan(restarted, { mode: 'tip' });
-
-    expect(recovered.map((cg) => cg.blockNumber)).toEqual([eventBlock]);
-    expect(registry.queryFilter.calls.map(([, lo, hi]: [unknown, number, number]) => [lo, hi])).toEqual([
-      [7_951, 9_950],
-      [9_951, 11_950],
-      [11_951, 13_200],
-    ]);
-    expect(await restarted.hasContextGraphRegistryScanWatermark()).toBe(false);
-  });
-
-  it('never aliases tip progress through a legacy role-less historical cursor store', async () => {
-    const store = new LegacyRolelessRegistryScanCursorStore();
-    await store.save({
-      chainId: 'evm:31337',
-      deploymentId: 'evm:31337:hub=0x0000000000000000000000000000000000000001',
-      registryAddress: REGISTRY,
-    }, 2_000);
-    store.loads.length = 0;
-    store.saves.length = 0;
-    const registry = makeRegistry();
-    const { adapter } = makeAdapter(registry, 10_000, {
-      contextGraphRegistryScanCursorStore: store,
-    });
-    registry.queryFilter.setImpl(async () => []);
-
-    await collectRegistryScan(adapter, { mode: 'tip' });
-
-    expect(registry.queryFilter.calls.map(([, lo, hi]: [unknown, number, number]) => [lo, hi])).toEqual([
-      [8_001, 10_000],
-    ]);
-    expect(store.loads).toEqual([]);
-    expect(store.saves).toEqual([]);
-    expect((adapter as any).contextGraphRegistryTipScanCursor.getCachedWatermark(REGISTRY)).toBe(10_001);
-
-    await expect(adapter.hasContextGraphRegistryScanWatermark()).resolves.toBe(true);
-    expect((adapter as any).contextGraphRegistryScanCursor.getCachedWatermark(REGISTRY)).toBe(2_000);
-    expect([...store.values.values()]).toEqual([2_000]);
-  });
-
-  it('retains the first tip lower bound in process when durable cursor persistence fails', async () => {
-    const warnSpy = vi.spyOn(console, 'warn').mockImplementation(() => {});
-    const store = {
-      cursorKeyVersion: 2 as const,
-      load: vi.fn(async () => undefined),
-      save: vi.fn(async () => {
-        throw new Error('tip cursor save failed');
-      }),
-    };
-    let head = 10_000;
-    let firstQuery = true;
-    const eventBlock = 10_001;
-    const registry = makeRegistry({
-      queryFilter: seam(async (_filter: unknown, lo: number, hi: number) => {
-        if (firstQuery) {
-          firstQuery = false;
-          throw new Error('first tip page unavailable');
-        }
-        return lo <= eventBlock && eventBlock <= hi
-          ? [{ topics: [], data: '0x01', blockNumber: eventBlock }]
-          : [];
-      }),
-    });
-    const { adapter, provider } = makeAdapter(registry, head, {
-      contextGraphRegistryScanCursorStore: store,
-    });
-    provider.getBlockNumber.setImpl(async () => head);
-
-    try {
-      await expect(collectRegistryScan(adapter, { mode: 'tip' }))
-        .rejects.toThrow('first tip page unavailable');
-      expect((adapter as any).contextGraphRegistryTipScanCursor.getCachedWatermark(REGISTRY))
-        .toBe(8_001);
-
-      head = 13_200;
-      registry.queryFilter.clear();
-      const recovered = await collectRegistryScan(adapter, { mode: 'tip' });
-
-      expect(recovered.map((cg) => cg.blockNumber)).toEqual([eventBlock]);
-      expect(registry.queryFilter.calls[0]?.slice(1)).toEqual([7_951, 9_950]);
-      expect((adapter as any).contextGraphRegistryScanCursor.getCachedWatermark(REGISTRY))
-        .toBeUndefined();
-    } finally {
-      warnSpy.mockRestore();
-    }
-  });
-
-  it('revisits a fetched tip page when local application never acknowledges it', async () => {
-    const store = new MemoryRegistryScanCursorStore();
-    const eventBlock = 9_999;
-    const registry = makeRegistry({
-      queryFilter: seam(async (_filter: unknown, lo: number, hi: number) =>
-        lo <= eventBlock && eventBlock <= hi
-          ? [{ topics: [], data: '0x01', blockNumber: eventBlock }]
-          : [],
-      ),
-    });
-    const { adapter } = makeAdapter(registry, 10_000, {
-      contextGraphRegistryScanCursorStore: store,
-    });
-    const iterator = adapter.scanContextGraphRegistryPages({ mode: 'tip' })[Symbol.asyncIterator]();
-
-    const fetched = await iterator.next();
-    expect(fetched.done).toBe(false);
-    expect(fetched.value?.contextGraphs.map((cg) => cg.blockNumber)).toEqual([eventBlock]);
-    await iterator.return?.();
-    expect(store.saves.map((save) => save.nextBlock)).toEqual([8_001]);
-
-    registry.queryFilter.clear();
-    const { adapter: restarted } = makeAdapter(registry, 10_000, {
-      contextGraphRegistryScanCursorStore: store,
-    });
-    const retried = await collectRegistryScan(restarted, { mode: 'tip' });
-
-    expect(retried.map((cg) => cg.blockNumber)).toEqual([eventBlock]);
-    expect(registry.queryFilter.calls.map(([, lo, hi]: [unknown, number, number]) => [lo, hi])).toEqual([
-      [7_951, 9_950],
-      [9_951, 10_000],
-    ]);
-    expect(await restarted.hasContextGraphRegistryScanWatermark()).toBe(false);
   });
 
   it('does not advance the incremental watermark when parsing a later page fails', async () => {
@@ -545,7 +169,7 @@ describe('EVMChainAdapter.listContextGraphsFromChain registry scan', () => {
     const registry = makeRegistry();
     const { adapter, provider } = makeAdapter(registry, 2_100, {
       cgRegistryScanPageSize: 1_000,
-      contextGraphRegistryScanCursorStore: store,
+      contextGraphRegistryScanCursorPersistence: roleAwareRegistryCursorPersistence(store),
     });
     registry.queryFilter.setImpl(async () => []);
 
@@ -618,7 +242,7 @@ describe('EVMChainAdapter.listContextGraphsFromChain registry scan', () => {
     });
     const { adapter } = makeAdapter(registry, 2_100, {
       cgRegistryScanPageSize: 1_000,
-      contextGraphRegistryScanCursorStore: store,
+      contextGraphRegistryScanCursorPersistence: roleAwareRegistryCursorPersistence(store),
     });
 
     const seeded = await adapter.listContextGraphsFromChain(undefined, {
@@ -711,7 +335,7 @@ describe('EVMChainAdapter.listContextGraphsFromChain registry scan', () => {
     const registry = makeRegistry();
     const { adapter } = makeAdapter(registry, 5_000, {
       cgRegistryScanPageSize: 1_000,
-      contextGraphRegistryScanCursorStore: store,
+      contextGraphRegistryScanCursorPersistence: roleAwareRegistryCursorPersistence(store),
     });
     registry.queryFilter.setImpl(async () => []);
 
@@ -729,7 +353,7 @@ describe('EVMChainAdapter.listContextGraphsFromChain registry scan', () => {
     const restartedRegistry = makeRegistry();
     const { adapter: restarted, provider } = makeAdapter(restartedRegistry, 5_000, {
       cgRegistryScanPageSize: 1_000,
-      contextGraphRegistryScanCursorStore: store,
+      contextGraphRegistryScanCursorPersistence: roleAwareRegistryCursorPersistence(store),
     });
     provider.getCode = seam(async () => {
       throw new Error('deploy block probing should not run with persisted cursor');
@@ -756,7 +380,7 @@ describe('EVMChainAdapter.listContextGraphsFromChain registry scan', () => {
     });
     const { adapter } = makeAdapter(registry, 2_100, {
       cgRegistryScanPageSize: 1_000,
-      contextGraphRegistryScanCursorStore: store,
+      contextGraphRegistryScanCursorPersistence: roleAwareRegistryCursorPersistence(store),
     });
 
     const iterator = adapter.scanContextGraphRegistryPages({
@@ -773,7 +397,7 @@ describe('EVMChainAdapter.listContextGraphsFromChain registry scan', () => {
     const restartedRegistry = makeRegistry();
     const { adapter: restarted } = makeAdapter(restartedRegistry, 2_100, {
       cgRegistryScanPageSize: 1_000,
-      contextGraphRegistryScanCursorStore: store,
+      contextGraphRegistryScanCursorPersistence: roleAwareRegistryCursorPersistence(store),
     });
     restartedRegistry.queryFilter.setImpl(async () => []);
 
@@ -868,7 +492,7 @@ describe('EVMChainAdapter.listContextGraphsFromChain registry scan', () => {
     const registry = makeRegistry();
     const { adapter, provider } = makeAdapter(registry, 5_000, {
       cgRegistryScanPageSize: 1_000,
-      contextGraphRegistryScanCursorStore: store,
+      contextGraphRegistryScanCursorPersistence: roleAwareRegistryCursorPersistence(store),
     });
     provider.getCode = seam(async () => {
       throw new Error('deploy block probing should not run with persisted cursor');
@@ -904,7 +528,7 @@ describe('EVMChainAdapter.listContextGraphsFromChain registry scan', () => {
     });
     const { adapter } = makeAdapter(registry, 3_500, {
       cgRegistryScanPageSize: 1_000,
-      contextGraphRegistryScanCursorStore: store,
+      contextGraphRegistryScanCursorPersistence: roleAwareRegistryCursorPersistence(store),
     });
 
     const results = await collectRegistryScan(adapter, {
@@ -925,7 +549,7 @@ describe('EVMChainAdapter.listContextGraphsFromChain registry scan', () => {
     const store = new MemoryRegistryScanCursorStore();
     const seedRegistry = makeRegistry();
     const { adapter: seedAdapter } = makeAdapter(seedRegistry, 2_100, {
-      contextGraphRegistryScanCursorStore: store,
+      contextGraphRegistryScanCursorPersistence: roleAwareRegistryCursorPersistence(store),
     });
     seedRegistry.queryFilter.setImpl(async () => []);
 
@@ -936,7 +560,7 @@ describe('EVMChainAdapter.listContextGraphsFromChain registry scan', () => {
 
     const publicRegistry = makeRegistry();
     const { adapter: publicAdapter } = makeAdapter(publicRegistry, 2_100, {
-      contextGraphRegistryScanCursorStore: store,
+      contextGraphRegistryScanCursorPersistence: roleAwareRegistryCursorPersistence(store),
     });
     publicRegistry.queryFilter.setImpl(async () => []);
 
@@ -967,7 +591,7 @@ describe('EVMChainAdapter.listContextGraphsFromChain registry scan', () => {
     const store = new MemoryRegistryScanCursorStore();
     const registry = makeRegistry();
     const { adapter } = makeAdapter(registry, 4_000, {
-      contextGraphRegistryScanCursorStore: store,
+      contextGraphRegistryScanCursorPersistence: roleAwareRegistryCursorPersistence(store),
     });
     registry.queryFilter.setImpl(async () => []);
