+++ packages/cli/test/daemon-snapshot-page-index-wiring.test.ts
@@ -107,7 +107,7 @@ function createFakeAgent() {
 function closeDashboardDbFromAgentCreateArg(createArg: any): void {
   const db =
     createArg?.chainEventCursorStore?.cursors?.db
-    ?? createArg?.contextGraphRegistryScanCursorStore?.cursors?.db;
+    ?? createArg?.contextGraphRegistryScanCursorPersistence?.store?.cursors?.db;
   db?.close?.();
 }
