+++ packages/cli/test/daemon-sync-agents-meta-wiring.test.ts
@@ -49,7 +49,7 @@ const { runDaemonInner } = await import('../src/daemon/lifecycle.js');
 function closeDashboardDbFromAgentCreateArg(createArg: any): void {
   const db =
     createArg?.chainEventCursorStore?.cursors?.db ??
-    createArg?.contextGraphRegistryScanCursorStore?.cursors?.db;
+    createArg?.contextGraphRegistryScanCursorPersistence?.store?.cursors?.db;
   db?.close?.();
 }
