+++ packages/cli/test/publisher-backfill-wiring-1828.test.ts
@@ -74,7 +74,7 @@ function createFakeServer() {
 function closeDashboardDbFromAgentCreateArg(createArg: any): void {
   const db =
     createArg?.chainEventCursorStore?.cursors?.db ??
-    createArg?.contextGraphRegistryScanCursorStore?.cursors?.db;
+    createArg?.contextGraphRegistryScanCursorPersistence?.store?.cursors?.db;
   db?.close?.();
 }
